+++ .idea/Udacity-Data-Engineer-NanoDegree.iml
@@ -1,7 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module type="PYTHON_MODULE" version="4">
   <component name="NewModuleRootManager">
-    <content url="file://$MODULE_DIR$" />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/p3_datawarehouse_aws_redshift/p3src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/p3_datawarehouse_aws_redshift" isTestSource="false" />
+    </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
