+++ p3_datawarehouse_aws_redshift/README.md
@@ -176,4 +176,4 @@ This table will:
 - [Using Dist Key and Sort Key (Flydata introduction)](https://www.flydata.com/blog/amazon-redshift-distkey-and-sortkey/)
 - [Copy from JSON (AWS documentation)](https://docs.aws.amazon.com/redshift/latest/dg/copy-usage_notes-copy-from-json.html)
 - [Rewritten query with the COPY command](https://aws.amazon.com/premiumsupport/knowledge-center/redshift-fix-copy-analyze-statupdate-off/)
-
+- [Redshift does not enforce primary key](http://www.sqlhaven.com/amazon-redshift-what-you-need-to-think-before-defining-primary-key/#:~:text=Redshift%20does%20not%20enforce%20these,exists%20in%20the%20referenced%20table)
