+++ p3_datawarehouse_aws_redshift/p3src/__init__.py
@@ -1,3 +1,3 @@
-from p3_datawarehouse_aws_redshift.p3src.createenv import create_tables_main
-from p3_datawarehouse_aws_redshift.p3src.createenv import create_cluster_main
-from p3_datawarehouse_aws_redshift.p3src.etl import etl_main
+from p3src.createenv import create_tables_main
+from p3src.createenv import create_cluster_main
+from p3src.etl import etl_main
