+++ p3_datawarehouse_aws_redshift/p3src/utils/aws.py
@@ -1,6 +1,7 @@
 import requests
 import boto3
 import pandas as pd
+import psycopg2
 
 def get_myip():
     """
@@ -47,3 +48,32 @@ def get_cluster_properties(config):
     x.loc['Endpoint_address'] = x.loc['Endpoint']['Address']
     x.loc['Role_arn'] = x.loc['IamRoles'][0]['IamRoleArn']
     return x
+
+
+def get_conn(config):
+    """
+
+    Args:
+        config: config file, see project readme for structure
+
+    Returns:
+        psycopg2.Connector
+    """
+    host = get_cluster_properties(config)['Endpoint_address']
+    DWH_DB = config.get("DB", "DB_NAME")
+    DWH_DB_USER = config.get("DB", "DB_USER")
+    DWH_DB_PASSWORD = config.get("DB", "DB_PASSWORD")
+    DWH_PORT = config.get("DB", "DB_PORT")
+    jdbcstring = "host={} dbname={} user={} password={} port={}".format(
+        host,
+        DWH_DB,
+        DWH_DB_USER,
+        DWH_DB_PASSWORD,
+        DWH_PORT
+    )
+    conn = psycopg2.connect(jdbcstring)
+    if conn.closed == 0:
+        print('Connection active: conn.closed == 0')
+    else:
+        print('Check connection')
+    return conn
