--- p3_datawarehouse_aws_redshift/project/exploration/psql.sh
@@ -1,2 +1,0 @@
-
-psql -h redshift-cluster-1.ccczdenythtf.eu-central-1.redshift.amazonaws.com -U dwhuser -p 5439 -d dwh
