--- p3_datawarehouse_aws_redshift/project/exploration/song_jsonpath
@@ -1,15 +1,0 @@
-{
-    "jsonpaths": [
-        "$['artist_id']",
-        "$['artist_latitude']",
-        "$['artist_location']",
-        "$['artist_longitude']",
-        "$['artist_name']",
-        "$['duration']",
-        "$['num_songs']",
-        "$['song_id']",
-        "$['title']",
-        "$['year']"
-
-    ]
-}
