--- p3_datawarehouse_aws_redshift/project/exploration/song_jsonpath.json
@@ -1,15 +1,0 @@
-{
-    "jsonpaths": [
-        "$['num_songs']",
-        "$['artist_id']",
-        "$['artist_latitude']",
-        "$['artist_longitude']",
-        "$['artist_location']",
-        "$['artist_name']",
-        "$['song_id']",
-        "$['title']",
-        "$['duration']",
-        "$['year']"
-
-    ]
-}
