+++ p3_datawarehouse_aws_redshift/project/routines/create_tables_main.py
@@ -1,6 +1,6 @@
-from p3_datawarehouse_aws_redshift.p3src import create_tables_main
+from p3src import create_tables_main
 import configparser
-# Input here the path to the admin config file
+# Input here the path to the db user config file
 config_path = '/Users/paulogier/81-GithubPackages/Udacity-Data-Engineer-NanoDegree/p3_datawarehouse_aws_redshift/project/config/dbuser_config.cfg'
 
 if __name__ == '__main__':
